[TASK] Extract mount point update logic from RecordMonitor
