+++ Classes/IndexQueue/RecordMonitor.php
@@ -25,14 +25,14 @@ namespace ApacheSolrForTypo3\Solr\IndexQueue;
  ***************************************************************/
 
 use ApacheSolrForTypo3\Solr\AbstractDataHandlerListener;
-use ApacheSolrForTypo3\Solr\Site;
+use ApacheSolrForTypo3\Solr\Domain\Index\Queue\RecordMonitor\Helper\MountPagesUpdater;
+use ApacheSolrForTypo3\Solr\GarbageCollector;
+use ApacheSolrForTypo3\Solr\IndexQueue\Queue;
 use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
-use ApacheSolrForTypo3\Solr\System\Page\Rootline;
 use ApacheSolrForTypo3\Solr\Util;
 use TYPO3\CMS\Backend\Utility\BackendUtility;
 use TYPO3\CMS\Core\DataHandling\DataHandler;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Frontend\Page\PageRepository;
 
 /**
  * A class that monitors changes to records so that the changed record gets
@@ -58,12 +58,22 @@ class RecordMonitor extends AbstractDataHandlerListener
     protected $indexQueue;
 
     /**
-     * Constructor
+     * Mount Page Updater
      *
+     * @var MountPagesUpdater
      */
-    public function __construct()
+    protected $mountPageUpdater;
+
+    /**
+     * RecordMonitor constructor.
+     *
+     * @param Queue|null $indexQueue
+     * @param MountPagesUpdater|null $mountPageUpdater
+     */
+    public function __construct(Queue $indexQueue = null, MountPagesUpdater $mountPageUpdater = null)
     {
-        $this->indexQueue = GeneralUtility::makeInstance('ApacheSolrForTypo3\\Solr\\IndexQueue\\Queue');
+        $this->indexQueue = is_null($indexQueue) ? GeneralUtility::makeInstance(Queue::class) : $indexQueue;
+        $this->mountPageUpdater = is_null($mountPageUpdater) ? GeneralUtility::makeInstance(MountPagesUpdater::class) : $mountPageUpdater;
     }
 
     /**
@@ -151,7 +161,7 @@ class RecordMonitor extends AbstractDataHandlerListener
                     if (!empty($record) && $this->isEnabledRecord($table,
                             $record)
                     ) {
-                        $this->updateMountPages($uid);
+                        $this->mountPageUpdater->update($uid);
 
                         $this->indexQueue->updateItem($table, $uid);
                     } else {
@@ -331,7 +341,7 @@ class RecordMonitor extends AbstractDataHandlerListener
     protected function doPagesPostUpdateOperations(array $fields, $recordUid)
     {
         $this->updateCanonicalPages($recordUid);
-        $this->updateMountPages($recordUid);
+        $this->mountPageUpdater->update($recordUid);
 
         if ($this->isRecursiveUpdateRequired($recordUid, $fields)) {
             $treePageIds = $this->getSubPageIds($recordUid);
@@ -391,7 +401,7 @@ class RecordMonitor extends AbstractDataHandlerListener
      */
     protected function removeFromIndexAndQueue($recordTable, $recordUid)
     {
-        $garbageCollector = GeneralUtility::makeInstance('ApacheSolrForTypo3\\Solr\\GarbageCollector');
+        $garbageCollector = GeneralUtility::makeInstance(GarbageCollector::class);
         $garbageCollector->collectGarbage($recordTable, $recordUid);
     }
 
@@ -439,92 +449,6 @@ class RecordMonitor extends AbstractDataHandlerListener
         foreach ($canonicalPages as $page) {
             $this->indexQueue->updateItem('pages', $page['uid']);
         }
-    }
-
-    // Mount Page Handling
-
-    /**
-     * Handles updates of the Index Queue in case a newly created or changed
-     * page is part of a tree that is mounted into a another site.
-     *
-     * @param int $pageId Page Id (uid).
-     */
-    protected function updateMountPages($pageId)
-    {
-        // get the root line of the page, every parent page could be a Mount Page source
-        /** @var $pageSelect PageRepository */
-        $pageSelect = GeneralUtility::makeInstance(PageRepository::class);
-        $rootLine = $pageSelect->getRootLine($pageId);
-
-        $destinationMountProperties = $this->getDestinationMountPropertiesByRootLine($rootLine);
-
-        if (!empty($destinationMountProperties)) {
-            foreach ($destinationMountProperties as $destinationMount) {
-                $this->addPageToMountingSiteIndexQueue($pageId,
-                    $destinationMount);
-            }
-        }
-    }
-
-    /**
-     * Finds Mount Pages that mount pages in a given root line.
-     *
-     * @param array $rootLineArray Root line of pages to check for usage as mount source
-     * @return array Array of pages found to be mounting pages from the root line.
-     */
-    protected function getDestinationMountPropertiesByRootLine(array $rootLineArray)
-    {
-        $mountPages = array();
-
-        $currentPage = array_shift($rootLineArray);
-        $currentPageUid = (int)$currentPage['uid'];
-
-
-        if (empty($rootLineArray) && $currentPageUid === 0) {
-            return $mountPages;
-        }
-
-            /** @var $rootLine Rootline */
-        $rootLine = GeneralUtility::makeInstance(Rootline::class, $rootLineArray);
-        $rootLineParentPageIds = $rootLine->getParentPageIds();
-
-        $pageQueryConditions = array();
-        if (!empty($rootLineParentPageIds)) {
-            $pageQueryConditions[] = '(mount_pid IN(' . implode(',', $rootLineParentPageIds) . '))';
-        }
-
-        if ($currentPageUid !== 0) {
-            $pageQueryConditions[] = '(mount_pid=' . $currentPageUid . ' AND mount_pid_ol=1)';
-        }
-        $pageQueryCondition = implode(' OR ', $pageQueryConditions);
-
-        $mountPages = $GLOBALS['TYPO3_DB']->exec_SELECTgetRows(
-            'uid, uid AS mountPageDestination, mount_pid AS mountPageSource, mount_pid_ol AS mountPageOverlayed',
-            'pages',
-            '(' . $pageQueryCondition . ') AND doktype = 7 AND no_search = 0'
-            . BackendUtility::deleteClause('pages')
-        );
-
-        return $mountPages;
-    }
-
-    /**
-     * Adds a page to the Index Queue of a site mounting the page.
-     *
-     * @param int $mountedPageId ID (uid) of the mounted page.
-     * @param array $mountProperties Array of mount point properties mountPageSource, mountPageDestination, and mountPageOverlayed
-     */
-    protected function addPageToMountingSiteIndexQueue(
-        $mountedPageId,
-        array $mountProperties
-    ) {
-        $mountingSite = Site::getSiteByPageId($mountProperties['mountPageDestination']);
-
-        $pageInitializer = GeneralUtility::makeInstance('ApacheSolrForTypo3\\Solr\\IndexQueue\\Initializer\\Page');
-        $pageInitializer->setSite($mountingSite);
-
-        $pageInitializer->initializeMountedPage($mountProperties,
-            $mountedPageId);
     }
 
     /**
